% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbsData.R
\name{filter_disjoint}
\alias{filter_disjoint}
\title{Удаляем образцы, по которым нельзя посчитать расстояния
с другими образцами. Начинаем с самых несвязанных}
\usage{
filter_disjoint(x, min_shared_cpg = 1)
}
\arguments{
\item{x}{rrbsMatrix object to filter}

\item{min}{threshold for min shared CpG count}
}
\description{
Удаляем образцы, по которым нельзя посчитать расстояния
с другими образцами. Начинаем с самых несвязанных
}
