% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbsData.R
\name{get_features}
\alias{get_features}
\title{Get features adjusted p-value (FDR adjusted) to distinct groups}
\usage{
get_features(x, sample_group, ...)
}
\arguments{
\item{x}{Data}

\item{sample_group}{Vector with group name for each sample}
}
\description{
Get features adjusted p-value (FDR adjusted) to distinct groups
}
