% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbsData.R
\name{loadRrbsData}
\alias{loadRrbsData}
\title{Loads .cov files of bismark format to S4 list like object.}
\usage{
loadRrbsData(ss, cpg, description, ...)
}
\arguments{
\item{ss}{A 'rrbsSampleList' object (list of 'rrbsSampleInfo') with sample properties (metadata).
Property named 'Cov_FileName' is mandatory.}

\item{cpg}{A}
}
\description{
Loads .cov files of bismark format to S4 list like object.
}
