% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbsData.R
\name{my.plot.hclust}
\alias{my.plot.hclust}
\title{plot hierarhical clustering}
\usage{
my.plot.hclust(hc, types, colors, main = "CpG methylation clustering",
  cex = 1, pch = 20)
}
\arguments{
\item{hc}{hclust object}

\item{types}{character vector with names by hc$labels}

\item{colors}{character vector with names by types}
}
\description{
plot hierarhical clustering
}
