% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrbsData.R
\name{site_res}
\alias{site_res}
\title{Resolves site with ambiguous IUPAC nucleotides to a vector of unambiguous sites.}
\usage{
site_res(site)
}
\arguments{
\item{site}{A character site as nucleotide seq}
}
\value{
A character vector of unambiguous nucleotides.
}
\description{
Resolves site with ambiguous IUPAC nucleotides to a vector of unambiguous sites.
}
\examples{
site_res('CHG')
}
